+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Dev
 
+### Enhancements
+
+* Add `A.Vector.find_index/2`
+
 ## v0.4.5 (2021-01-31)
 
 ### Enhancements